v12. + Single Responsibility.
