--- hw06-ATM/src/main/java/atm/Balanceable.java
@@ -1,5 +1,0 @@
-package atm;
-
-public interface Balanceable {
-    Integer getBalance();
-}
