+++ hw06-ATM/src/main/java/atm/atm/ATM.java
@@ -1,11 +1,13 @@
 package atm.atm;
 
 import atm.Nominal;
+import atm.cell.Cell;
+
 import java.util.List;
 
 public interface ATM {
     boolean putCash( Nominal cash);
     List<Nominal> getCash(Integer sum);
-    Integer getBalance();
+    List<Cell> getAtmStorage();
 }
