+++ hw06-ATM/src/main/java/atm/atm/ATMService.java
@@ -1,0 +1,9 @@
+package atm.atm;
+
+import atm.cell.Cell;
+import java.util.List;
+
+public interface ATMService {
+
+    Integer getBalance();
+}
